\name{eximg}
\alias{eximg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Utility function}
\description{
Exports sample image files to R temporary directory and returns a path to the sample image files. This function is only used in the example of \code{\link{run.ij}}.
}
\usage{
eximg()
}
%- maybe also 'usage' for other objects documented here.
\value{
Jpeg files to R temprary directory and a path to the sample image files.
}

\author{
Masatoshi Katabuchi <mattocci27@gmail.com>}

%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
\code{\link{run.ij}} 
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line


