\name{LeafArea-package}
\alias{LeafArea-package}
\alias{LeafArea}
\docType{package}
\title{
LeafArea: A rapid digital image analysis for leaf area
}
\description{
The package LeafArea allows one to conveniently run IMAGEJ software within R. Functions in this package analyze multiple leaf images in the target directory and generate multiple data frame objects that include leaf area from each leaf image, and then process and combine these data frame objectes into a single data frame in a format that is convenient for subsequent analyses. The package provides a user-friendly, automated tool for measuring leaf area from digital images. For more information on ImageJ, see the ImageJ User Guide, which is available \url{http://imagej.nih.gov/ij/}.
}
\details{
\tabular{ll}{
Package: \tab LeafArea\cr
Type: \tab Package\cr
Version: \tab 0.0.5\cr
Date: \tab 2015-02-02\cr
License: \tab GPL-2\cr
}
}
\author{
Masatoshi Katabuchi <mattocci27@gmail.com>
}
\references{
Rasband, W.S., ImageJ, U. S. National Institutes of Health, Bethesda, Maryland, USA,  \url{http://imagej.nih.gov/ij/}, 1997-2014.\cr

Schneider, C.A., Rasband, W.S., Eliceiri, K.W. "NIH Image to ImageJ: 25 years of image analysis". Nature Methods 9, 671-675, 2012.\cr

Abramoff, M.D., Magalhaes, P.J., Ram, S.J. "Image Processing with ImageJ". Biophotonics International, volume 11, issue 7, pp. 36-42, 2004. (This article is available \url{http://dspace.library.uu.nl/handle/1874/204900}.)\cr
}
\keyword{ package }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line

