\name{leafdata}
\alias{leafdata}
\docType{data}
\title{ LeafArea default data }
\description{
  Analyzed leaf area data for example
}
\usage{data(leafdata)}
\format{
    A list with six data frames, A1-01.jpeg.txt, A1-02.jpeg.txt, A123-01.jpeg.txt, A123-02.jpeg.txt, A2.jpeg.txt and A300-1.jpeg.txt. The first column is the identity of image, and the second column is the are of the image (cm2).
}

\keyword{datasets}
