% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment_mice.R
\name{outlier_treatment_mice}
\alias{outlier_treatment_mice}
\title{Outlier Treatment using MICE Imputation}
\usage{
outlier_treatment_mice(data_frame, multiplier = 1.5)
}
\arguments{
\item{data_frame}{A data frame containing numeric variables where outliers will be treated.}

\item{multiplier}{The multiplier to determine the threshold for identifying outliers. Default is 1.5.}
}
\value{
A data frame with outliers removed and missing values imputed using the MICE algorithm.
}
\description{
This function identifies and removes outliers from a given data frame and then performs
missing value imputation using the MICE (Multivariate Imputation by Chained Equations) algorithm.
}
\details{
The function first identifies outliers using the IQR method and replaces them with NA.
Then, MICE imputation is applied to fill in missing values.
}
\examples{
# Example with a custom data frame
data <- data.frame(
  Culture = c("S. quitoense", "S. quitoense", "S. quitoense", "S. quitoense",
              "S. quitoense", "S. quitoense", "S. betaceum", "S. betaceum",
              "S. betaceum", "S. betaceum", "S. betaceum", "S. betaceum"),
  LengthLeaf = c(25, 27, 200, 56, NA, 56, 25, 27, 28, 56, 56, 56),
  WidthLeaf = c(45, 56, 57, 53, 54, NA, 57, 56, 57, 53, 54, 55)
)
df <- outlier_treatment_mice(data)
df

}
\references{
Van Buuren, S. (2018). Flexible Imputation of Missing Data. Chapman & Hall/CRC.
}
\seealso{
\code{\link{boxplot}} for creating boxplots.
\code{\link[mice]{mice}} for the MICE algorithm.
}
\author{
Velasquez-Vasconez
}
\keyword{imputation}
\keyword{mice}
\keyword{outlier}
\keyword{treatment}
