\name{readtext.ij}
\alias{readtext.ij}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
File management
}
\description{
Read multiple tab-delimited text files with a leaf area value (one text file for each original JPEG image file) generated by ImageJ.}

\usage{
readtext.ij(path)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path}{
Path to the target directory
}
}

\value{
A list of data frame of leaf area for each image
    \item{area}{
    Leaf area of the image (cm2)}
}

\author{
Masatoshi Katabuchi <mattocci27@gmail.com>}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{run.ij}}
\code{\link{resmerge.ij}}
 }
\examples{
#prepare example files
data(leafdata)
tf <- paste(tempdir(),"/",sep="")
for (i in 1:7){
	write.table(leafdata[[i]],paste(tf,names(leafdata)[i],sep=""),sep="\t")
}

#list of files
list.files(tf)

readtext.ij(tf)

unlink(list.files(tf))
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line

