% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eximg.r
\name{eximg}
\alias{eximg}
\title{Utility function}
\usage{
eximg()
}
\value{
Jpeg files to the R temporary directory and a path to the sample image files.
}
\description{
Exports sample image files to R temporary directory and returns a path to the sample image files. This function is only used in the example of \code{\link{run.ij}}.
}
\seealso{
\code{\link{run.ij}} which this function wraps
}
\author{
Masatoshi Katabuchi \email{mattocci27@gmail.com}
}
