% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.ij.R
\name{find.ij}
\alias{find.ij}
\title{Checking a path to ImageJ}
\usage{
find.ij(ostype = .Platform$OS.type)
}
\arguments{
\item{ostype}{the Operating system types.}
}
\value{
A path for ImageJ application. On unix this will always be 
"/Applications/ImageJ/". On Windows this will always be "C:\\Program Files\\ImageJ\\".
}
\description{
Check if ImageJ is installed in the correct directory.
}
\details{
When ImageJ, including ij.jar and Java, is not installed in the common location, users need to specify the path to ImageJ in \code{\link{run.ij}}.
}
\author{
Masatoshi Katabuchi \email{mattocci27@gmail.com}
}
