\name{leafdata}
\alias{leafdata}
\docType{data}
\title{ LeafArea default data }
\description{
  Analazed leafarea data for example
}
\usage{data(leafdata)}
\format{
    A list with six data frames, A1-01.jpeg.txt, A1-02.jpeg.txt, A123-01.jpeg.txt, A123-02.jpeg.txt, A2.jpeg.txt and A300-1.jpeg.txt. The first colmun is the identiy of image, and the second colmun is the are of the image (cm2).
}

\keyword{datasets}
