% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafdata.r
\docType{data}
\name{leafdata}
\alias{leafdata}
\title{LeafArea default data}
\format{
A list with six data frames, A1-01.jpeg.txt, A1-02.jpeg.txt, A123-01.jpeg.txt, A123-02.jpeg.txt, A2.jpeg.txt and A300-1.jpeg.txt. The first column is the identity of image, the second column is the area of the image (cm2) and the thrid column is the perimeter of the image (cm).
}
\usage{
leafdata
}
\description{
Analyzed leaf area data for example
}
\keyword{data}
